Remove WEBHOOK_URL variable
The WEBHOOK_URL variable in .env currently points to the domain of A&M channel 1.

Given that this particular domain functions as a "mailbox" for network requests to channel 1, the existence of the variable causes all messages that were intended for channels 2 and 3 to be sent to channel 1. For instance, if a user invokes the /start command in channel 2, the welcome message will be sent to the user via channel 1.

Let's remove the WEBHOOK_URL variable from .env and set each channel's WEBHOOK_URL config var to their own Heroku domains.